Maint/build ~ fix/Makefile.win ~ improve OSID selection heuristic
- require '.exe' extension for ${SHELL} as additional criteria before declaring OSID as "win"
- avoid incorrectly identifying MSYS/MSYS2 as Windows

## [why]

`make` may, incorrectly, choose a *nix-like shell (bash/sh/...) on a Windows host if that
shell is found on the PATH. And, recent versions of Win10, have `bash` on the path at
"%SystemRoot%\System32\bash.exe", at least for those with one or more WSL installations.
This requires a ${SHELL} reset to ${COMSPEC} (usually `cmd.exe`) which is commonly
triggered on the value of ${OS} being "Windows_NT".

MSYS/MSYS2 are *nix-like hosts and should be treated as *nix. But, determining whether
`make` is being executed under an MSYS/MSYS2 *nix-like shell is not straightforward. The
common method of using the value of ${OS} (as "Windows_NT") to determine non-*nix
execution will likely fail as the MSYS/MSYS2 shell is run as a Windows process and
can/will inherit that ${OS} value. ${MSYSTEM}, set by MSYS/MSYS2, is of no help as it
may be set by the installer and have an MSYS/MSYS2-type value prior to executing an
MSYS/MSYS2 shell.

So, as an additional heuristic to determine a correct ${SHELL} value, the initial
${SHELL} selection made initially by `make` must end in ".exe" along with ${OS} equal to
"Windows_NT" to declare OSID as "win". Otherwise, "nix" is used as the OSID fallback.